requirements.txt
moviepy==2.1.1
streamlit==1.43.1
edge_tts==6.1.19
fastapi==0.115.6
uvicorn==0.32.1
openai==1.56.1
faster-whisper==1.1.1
loguru==0.7.2
google-generativeai==0.8.4
dashscope==1.20.14
g4f==0.3.8.1
azure-cognitiveservices-speech==1.41.1
redis==5.2.0
python-multipart==0.0.19
streamlit-authenticator==0.4.1
PyYAML==6.0.2
cryptography==44.0.2
sqlalchemy==2.0.39
pymysql==1.1.1
pydantic==2.10.6
pytz==2025.1
fastapi-pagination==0.12.34
jinja2==3.1.6
pytest==8.3.5
python-dotenv
supabase==2.15.0
python-dotenv==1.1.0